+++ ChangeLog.md
@@ -13,6 +13,7 @@ Other enhancements:
 * `--reconfigure` flag [#914](https://github.com/commercialhaskell/stack/issues/914) [#946](https://github.com/commercialhaskell/stack/issues/946)
 * Cached data is written with a checksum of its structure [#889](https://github.com/commercialhaskell/stack/issues/889)
 * Fully removed `--optimizations` flag
+* Added `--cabal-verbose` flag
 
 Bug fixes: